requirements.txt
kafka_python==2.0.2
python-dotenv==1.0.1
Requests==2.32.3
